+++ R/svd.R
@@ -8,7 +8,7 @@
 #' bidiagonal divide and conquer strategy (BDC) SVD, and (2) two-sided Jacobi
 #' SVD for small matrices (<16) and high accuracy.
 #'
-#' @param a Square numeric matrix with the coefficients of the linear system.
+#' @param a Numeric matrix.
 #' @param type Character scalar. Whether to use BDC or Jacobi SVD.
 #' @param vectors Logical scalar indicating whether singular vectors should be
 #' computed and returned.
@@ -16,12 +16,12 @@
 #' returned in thinned or full format.
 #'
 #' @return Solves the generalised eigenproblem and returns a list with
-#' singular values in the \code{"d"} slot and, if requested, singular vectors
-#' in the slots \code{"u"} and \code{"v"}.
+#' singular values in the \code{"d"} component and, if requested, singular
+#' vectors in the components \code{"u"} and \code{"v"}.
 #'
 #' @export
 #' @examples
-#' \dontshow{set.seed(42)}
+#' set.seed(42)
 #' # Compute singular values and vectors using BDC
 #' A <- matrix(rnorm(9), nrow = 3, ncol = 3)
 #' sv <- svd2(A)
@@ -41,16 +41,16 @@ svd2 <- function(a, type = c("BDC", "Jacobi"), vectors = TRUE, thin = TRUE) {
   # Checks -----
   type <- match.arg(type)
 
-  if(!is.matrix(a)) {stop("Please provide 'a' as matrix.")}
+  if(!is.matrix(a)) {stop("Please provide 'a' as matrix or sparse 'dgCMatrix'.")}
   vectors <- isTRUE(vectors)
   thin <- isTRUE(thin)
 
-  type <- if(vectors) {if(thin) {0} else {1}} else {2}
+  thin_full <- if(vectors) {if(thin) {0} else {1}} else {2}
 
   # Execute -----
   if(type == "Jacobi") {
-    return(svd_J(a, type = type))
+    return(svd_J(a, type = thin_full))
   } else {
-    return(svd_BDC(a, type = type))
+    return(svd_BDC(a, type = thin_full))
   }
 }